% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selector.R
\name{cal_AUROC}
\alias{cal_AUROC}
\title{cal_AUROC}
\usage{
cal_AUROC(fea_, target_)
}
\arguments{
\item{fea_}{Features for calculation}

\item{target_}{Outcome variable used for calculation, which should be a binary variable for OR and AUROC}
}
\description{
cal_AUROC
}
