% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selector.R
\name{cal_cor}
\alias{cal_cor}
\title{cal_cor}
\usage{
cal_cor(fea_, target_, method_ = "pearson", pvalue_ = F, ...)
}
\arguments{
\item{fea_}{Features for calculation}

\item{target_}{Outcome variable used for calculation,
according to different calculation methods, data types can be integers, continuous numeric, or factors.}

\item{method_}{Evaluation methods for correlation, including pearson spearman, and kendall correlation coefficient}

\item{pvalue_}{If the test would be executed. Output the p-value if T.}

\item{...}{Further arguments to be passed to or from methods.}
}
\description{
cal_cor
}
