% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selector.R
\name{cal_HR}
\alias{cal_HR}
\title{cal_HR}
\usage{
cal_HR(fea_, time_, event_)
}
\arguments{
\item{fea_}{Features for calculation}

\item{time_}{Time when the event occurred or the experiment ended}

\item{event_}{Whether the termination event occurs or not}
}
\description{
cal_HR
}
