% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selector.R
\name{cal_variance}
\alias{cal_variance}
\title{cal_variance}
\usage{
cal_variance(fea_, method_ = c("variance", "kw", "wilcoxon"), ...)
}
\arguments{
\item{fea_}{Features for calculation}

\item{method_}{Evaluation methods for dispersion, including variance, p-values of Kruskal or Wilcoxon rank sum test.
If using Kruskal or Wilcoxon rank sum test, an additional grouping variable needs to be provided.}
}
\description{
cal_variance
}
