% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_enrich.R
\name{enrichGraph}
\alias{enrichGraph}
\title{enrichGraph}
\usage{
enrichGraph(
  inputGenes_,
  db_,
  pathways_ = "all",
  bgNet_ = "reactome",
  thres_ = 0.05
)
}
\arguments{
\item{inputGenes_}{A string of genes you want to enrich.}

\item{db_}{Data frame of the enriched item, the first column is term and the second column is gene.}

\item{pathways_}{The specific path term used to perform enrichment, pass \code{all} to enrich all terms.}

\item{bgNet_}{Background network type. Currently, prior gene-gene interactions from Reactome and STRING databases are included.
You can specify to use either 'reactome' or 'string' by passing the respective argument.
If you wish to use custom gene-gene interaction pairs, please pass a dataframe containing two columns of genes.}

\item{thres_}{FDR threshold of the output result.}
}
\description{
enrichGraph
}
\examples{
db = read_gmt('material/h.all.v2023.2.Hs.symbols.gmt')
result = enrichGraph(geneString, db, 'all')
}
