% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_enrich.R
\name{enrichGraphInner}
\alias{enrichGraphInner}
\title{enrichGraphInner}
\usage{
enrichGraphInner(inputGenes_, geneSet_, bgNet_, bgGeneNum_)
}
\description{
Used to perform enrichment of a single gene set in a prior gene interaction pair.
}
\keyword{internal}
