% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pairwise.R
\name{pair.wise}
\alias{pair.wise}
\title{pair.wise}
\usage{
pair.wise(mat_, net_, nThreads_ = 1)
}
\arguments{
\item{mat_}{Expression matrix, which is expected to have Gene Symbol represented by rownames and samples represented by colnames.}

\item{net_}{Background network}

\item{nThreads_}{Threads to use for transformations, the recommended number of  is between 3 and 6.}
}
\description{
Used to output a binary matrix according to the expression level of the prior gene pairs
}
\keyword{internal}
