% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deltarank.R
\name{rank.matrix}
\alias{rank.matrix}
\title{rank.matrix}
\usage{
rank.matrix(mat_)
}
\arguments{
\item{mat_}{Expression matrix, which is expected to have Gene Symbol represented by rownames and samples represented by colnames.}
}
\description{
Used to convert the expression matrix to the rank matrix of the expression
}
\keyword{internal}
