% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_enrich.R
\name{read_gmt}
\alias{read_gmt}
\title{read_gmt}
\usage{
read_gmt(gmtfile_, filteron_ = NULL)
}
\arguments{
\item{gmtfile_}{Path to gmt formatted file.}

\item{filteron_}{Filter terms without keywords like \code{GOBP}.}
}
\description{
Used to read in gmt file downloaded form [MSigDB](https://www.gsea-msigdb.org/gsea/index.jsp)
}
\examples{
db = read_gmt('yourpath/h.all.v2023.2.Hs.symbols.gmt')
}
