% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{transGI}
\alias{transGI}
\title{transGI}
\usage{
transGI(
  testMat_,
  method_ = "deltarank",
  bgNet_ = "reactome",
  nThreads_ = 1,
  maskMat_ = NULL,
  controlMat_ = NULL
)
}
\arguments{
\item{testMat_}{Expression matrix, which is expected to have Gene Symbol represented by rownames and samples represented by colnames.}

\item{method_}{Transformation method, currently consisting of either DeltaRank or Pairwise.
For detailed information, please refer to \code{\link{delta.rank}} or \code{\link{pair.wise}}.}

\item{bgNet_}{Background network type. Currently, prior gene-gene interactions from Reactome and STRING databases are included.
You can specify to use either 'reactome' or 'string' by passing the respective argument.
If you wish to use custom gene-gene interaction pairs, please pass a dataframe containing two columns of genes.}

\item{nThreads_}{Threads to use for transformations, the recommended number of  is between 3 and 6.}

\item{maskMat_}{A binary matrix, with the same dimension as \code{testMat_}, is integrated-term-by-term with \code{testMat_},
which means that the values of the corresponding positions in \code{testMat_} are penalized to the minimum value for function-based integration.}

\item{controlMat_}{Expression matrix with the same dimensions as \code{testMat_}, usually normal tissue.
The mean value of expression in the matrix and the corresponding transformation result are subtracted from the final result
as the level of gene expression perturbation in the control group.}
}
\description{
Nonparametric transformation method of transcriptome based on prior gene interactions
}
\examples{
testMat_ = system.file('extdata', 'inputMatTest.csv', package = 'transGI') |>
  read.csv(row.names = 'symbol') |>
  as.matrix()
result = transGI(testMat_, 'deltarank', 'reactome')
}
