% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{visNet}
\alias{visNet}
\title{visNet}
\usage{
visNet(inputGenes_, bgNet_ = "reactome", degree_ = 3)
}
\arguments{
\item{inputGenes_}{A string of genes you want to enrich.}

\item{bgNet_}{Background network type. Currently, prior gene-gene interactions from Reactome and STRING databases are included.
You can specify to use either 'reactome' or 'string' by passing the respective argument.
If you wish to use custom gene-gene interaction pairs, please pass a dataframe containing two columns of genes.}

\item{degree_}{An integer. Label nodes with node degree more than it.}
}
\description{
visNet
}
